[SCB-2278]in safe mode should not query instances
