+++ service-registry/registry-service-center/src/main/java/org/apache/servicecomb/serviceregistry/event/RecoveryEvent.java
@@ -14,14 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.servicecomb.serviceregistry.event;
 
-package org.apache.servicecomb.registry.api.event.task;
-
-/**
- * Service Center Heart Beat success event
- *
- */
-public class HeartbeatSuccEvent {
+public class RecoveryEvent {
 
 }
-
