+++ service-registry/registry-service-center/src/main/java/org/apache/servicecomb/serviceregistry/registry/RemoteServiceRegistry.java
@@ -21,16 +21,20 @@ import java.util.List;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import org.apache.commons.configuration.Configuration;
 import org.apache.servicecomb.foundation.common.concurrency.SuppressedRunnableWrapper;
 import org.apache.servicecomb.foundation.common.utils.SPIServiceUtils;
 import org.apache.servicecomb.serviceregistry.client.ServiceRegistryClient;
 import org.apache.servicecomb.serviceregistry.config.ServiceRegistryConfig;
+import org.apache.servicecomb.serviceregistry.event.HeartbeatFailEvent;
+import org.apache.servicecomb.serviceregistry.event.HeartbeatSuccEvent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.eventbus.EventBus;
+import com.google.common.eventbus.Subscribe;
 
 public class RemoteServiceRegistry extends AbstractServiceRegistry {
   private static final Logger LOGGER = LoggerFactory.getLogger(RemoteServiceRegistry.class);
@@ -39,6 +43,8 @@ public class RemoteServiceRegistry extends AbstractServiceRegistry {
 
   private List<ServiceRegistryTaskInitializer> taskInitializers = SPIServiceUtils
       .getOrLoadSortedService(ServiceRegistryTaskInitializer.class);
+
+  private AtomicBoolean heartBeatStatus = new AtomicBoolean(true);
 
   public RemoteServiceRegistry(EventBus eventBus, ServiceRegistryConfig serviceRegistryConfig,
       Configuration configuration) {
@@ -82,6 +88,10 @@ public class RemoteServiceRegistry extends AbstractServiceRegistry {
 
     taskPool.scheduleAtFixedRate(
         new SuppressedRunnableWrapper(() -> {
+          if (!heartBeatStatus.get()) {
+            LOGGER.warn("heart beat currently not success, pause for pulling instance.");
+            return;
+          }
           serviceRegistryCache.refreshCache();
         }),
         serviceRegistryConfig.getInstancePullInterval(),
@@ -95,5 +105,18 @@ public class RemoteServiceRegistry extends AbstractServiceRegistry {
 
   public ScheduledThreadPoolExecutor getTaskPool() {
     return this.taskPool;
+  }
+
+  @Subscribe
+  public void onHeartbeatSuccEvent(HeartbeatSuccEvent heartbeatSuccEvent) {
+    // 可以考虑多等待一个心跳周期，这样的好处是尽可能避免provider滞后于consumer注册的情况，consumer访问provider失败
+    // 但是这样的坏处是当服务中心两个实例，一个可用，另外一个不可用的情况下，长期处于不查询实例状态。 目前按照一次的实际效果
+    // 还是挺好的，也是和版本一直的实现一致。
+    heartBeatStatus.set(true);
+  }
+
+  @Subscribe
+  public void onHeartbeatFailEvent(HeartbeatFailEvent heartbeatFailEvent) {
+    heartBeatStatus.set(false);
   }
 }
