+++ service-registry/registry-service-center/src/main/java/org/apache/servicecomb/serviceregistry/task/MicroserviceWatchTask.java
@@ -21,8 +21,8 @@ import org.apache.servicecomb.registry.api.registry.Microservice;
 import org.apache.servicecomb.registry.api.event.MicroserviceInstanceChangedEvent;
 import org.apache.servicecomb.serviceregistry.client.ServiceRegistryClient;
 import org.apache.servicecomb.serviceregistry.config.ServiceRegistryConfig;
-import org.apache.servicecomb.registry.api.event.task.ExceptionEvent;
-import org.apache.servicecomb.registry.api.event.task.RecoveryEvent;
+import org.apache.servicecomb.serviceregistry.event.ExceptionEvent;
+import org.apache.servicecomb.serviceregistry.event.RecoveryEvent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
